rust:macos-build:add libmacros.dylib to MRPROPER_FILES
libmacros crate file is part of the build infarstructure and as such is
exempted from beeing remove by the clean target.

On macOS the crate file is named libmacros.dylib so add this to the list
of files that are cleaned by the mrproper target.

Apparently there is no modular way to add file to the mrproper target so
the only way is to add the file name to MRPROPER_FILES variable?

Signed-off-by: Alexis Deruelle <alexis.deruelle@gmail.com>