rust:macos-build:compute proc_macro crate prefix and suffix
On non-Linux platform proc_macro crates files have different extension
and specifically on macOS it is .dylib instead of .so

What we do here is calling Rust compiler with print option to univocally
retrieve naming convention of proc macro crates and then computing
prefix and suffix for such crates on the host platform.

This is slightly overkill but why not?

Signed-off-by: Alexis Deruelle <alexis.deruelle@gmail.com>